Split upgrade card to separate file.
